the departing deer have come of age simplest stuff first - ecs: slim stylish, basic, no frills. probably works. v nice. I'm gonna put it in its own repos later. there are 3 components, the storage, the handle manager and the view. the storage is a sparse array look up from the entity to the components. the handle manager creates unique handles and handles their destruction somewhat. *it does not erase the components associated with the entity*. the view simply stores a pointer to the array and contains and iterator pair for forward iteration over the array.
the vulkan stuff:
 Ive begun seperating off functionality from the original graphics object. ive got it so far into: Context - contains the instance and device
Window - contains the glfw window, surface, swapchain and presentation renderpass.
Renderer - contains the render frames, shader programs and semaphores and meshes[NOT DONE]
Scene - probably inherits the Renderer [NOT DONE]

in the graphics I have a working version of my current understanding of vulkan to test things and ensure i have something to copy. this is almost the same as last version.

it now resizes without recompiling the VkPipeline. this done through a built in feature which passes the "extent" and "scissors" to a pipeline before drawing.