rocks in my path. everything is gone. everything. voluntarily. all the vulkan v2 stuff goes.
theres a mess in main.cpp where I'm playing with the ecs all together. its looking like a big mess but progress is being made. the pool has largely stayed the same the main difference being the reorder function which replaces the sort. its pretty cool. it takes a function which reorders the entities and then runs a loop to swap the components until their in the matching order. (it was a nightmare to figure out).

registry is now type erased which is very fancy. not sure its good or performance but oh well. this is done with a hacky adaptation of std::unique_ptr, each pool is a stored in a std::unique_ptr with a delete which basically just inherits a lambda function which calls a delete on the typed template. still a bit confused that it works. the type lookup uses std::type_index which seems to work and i just use it in an unordered_map

the view now more specifically gets and includes different types. I still have to mess around with the registry function. but the view now stores a reference for each pool it uses. the plan is that each non const pool reference will use a mutex lock which will means to access the pool you have to wait for the view to be unlocked. this is all part of the multithreaded design stuff that i got goin on. still working on it clearly. I need a good way to enable and disable the mutex while the view is iterating. its which is more annoying than you'd think.

the dispatcher now takes a function pointer, its now in the pool tho its currently not dispatched ever. its kinda a problem for the multithreaded part.

multithreaded memory access can be done in 2 ways the granular way and the broad strokes way. locking per iteration is a broad strokes approach, but a event dispatcher probably requires the granular approach, theres nothing wrong with doing both but the dispatcher will probably have to wait until the iterator is finished which isnt necessarily very fast.

oh also handle manager is gone, it too needs have the same thread memory access management as the pools. so ima do that.

this will probably be implemented through a wrapper around a reference to the type, however you lose the interface if you do that so i gotta think of something else.