flirting is when 2 threads access the same resource
std::shared_mutex is what I want
the pool has a std::shared_mutex and a const instance holds a std::shared_lock while a non const instance holds a std::unique_lock.

there is currently a potential risk of threads trying to instance a pool at the same time. this will break everything. This is resolved if:
all pools are already instanced
a shared_mutex and a unique_lock encapsulated the functions

another issue. Theres a deadlock waiting to happen. if 2 views are created at the same time in reverse order, the first will wait on the second and the second will wait on the first.

update:
added Pool Policies 1 for a typed and 1 for a type erased. the typed stores all the pools in a tuple but they must be declared beforehand and the type erased uses an unordered_map. both have solved the dead lock issue above. typed sorts the pools and initiates them in a consistent order and the type erased uses a new mutex and lock for the duration of creating the pool. Its still a problem, the user can call pool() individually.

I removed the Dispatcher class. The whole event system relies on any pools it uses being unlocked. It would also deadlock when a thread requested a pool twice. overall it just wasnt very well thought through. It may return as the handle manager suffers the same issue so if I can think of a elegant work around I'll try and implement it. I think other things take priority though especially as event driven programming is the opposite of cache friendly.

Possible ToDo:
refactor the view as a pipeline -> logically this works much better. you can inherit the pipeline and do your processing stuff. A registry will initiated a pipeline each time its run, this will involve getting all the locks etc

new update:
forget all that. no untyped registry -> i dont like.
access management is now managed by the pipeline class making it entirely optional to implement. its much cleaner now too. iteration over tuples is much cleaner now also, a lambda is template lambda is inside a fold expression. it shouldnt have a significant performance impact for anything other than compile time.

View filters are back baybee.

not entirely sure pipeline should be used how it currently is used. it needs further work. objects created from a pipeline are invalidated when the pipeline is deleted. This is good we like. I might remove encapsulate the functions in the registry from the user to enforce access through the pipeline? maybe.

multithreading is cool

ToDo:
-HandleManager
-Finalize ECS