jacked up new config setup to add for settings
tested the ECS with imported systems from tentillium -> lead to changes all over:
More control over how a view iterates over the pools, can specify which pool to use as the orderby type. must also now specify that Entity is requested.

incorporated the handlemanager again. there is a new logical inheritance a Storage and a HandleManager are both types of Pool and both inherit AccessManager which contains the mutex and locking functions thus there are 2 types of pool a Entity Pool and a Component Pool. Both can be used as the orderby type

all types are now back to being nested classes of registry(other than registry). I changed this because although error messages are hideous, it's tidier code.

The Iterator in View uses the vector iterator, same with the new iterator in pool and handlemanager

fairly sure the ECS file was renamed to Scene -> dont remember doing it

After testing the ECS stuff I thought I should comment on a few things: iterating over the handle manager does not remove the invalid entities. It's not great but it might discourage people trying it. it doesnt really make a whole lot of sense anyway because an Iterator is supposed to point/reference a position in memory where as this is just a range.

the Vulkan stuff hasnt progressed too far. mostly refamiliarising myself with whats going on made a couple of changes but nothing major

started on the path of creating a structured Graphics class to replace the 1000 line mess. So I've got instance and device setup again, the vulkan handle encapsulation and config files.

BLAAARGH
include glm from Vulkan SDK not local