I'm really growing as a person. My foot fungi thriving.
bug fixes pretty much the last commit where I merged, the merge broke the thing I just did so... good job there.

I changed the HandleManager Iterator. so it now only gives you valid enities. I dont love it but it is what it is. Pipeline.

Renamed the types in the view function, It used to be sql like because I was trying to be fancy. I have ceased being fancy.

in the Storage class fixed the constexpr guards that prevent the component being used when its not being stored.

Oh also say hello to Swapchain, I think it broke the commit last time. not really sure.