might act on throws of whimsy and free abandon
COMP
renamed transform WorldMatrix and LocalMatrix to World and Local

ECS
updated the handle manager - it nows uses a bidrectional and forward linked list inside a sparse table. the bidirectional linked list allows fast deletion and the sparse array allows fast lookup times for alive entities. when an entity is destroyed it uses the forward linked list to store the dead entities. when a entity is created it first checks the dead list and if empty will append to the sparse array. to validate if an array is alive the handle manager uses the prev ptr and checks if it isnt pointing to anything. the linked list is using ptr diffference so vector resizing causes no issues.

moved the storage and handlemanager out of the registry so they are no longer subclasses

CORE
moved the config.h to the core folder
removed the graphics namespace